*refactoring of ConvertHardSigmoidNode.
Move ConvertHardSigmoidNode from OnnxRuntime.h into ConvertHardSigmoidNode.cpp.
